Added some atom functions

+++ src/ewmh.h
@@ -32,8 +32,8 @@ void ewmh_init(Display *, Window);
 void ewmh_root_client_list_add(Display *, Window, Window);
 void ewmh_root_client_list_clear(Display *, Window);
 void ewmh_root_set_active_window(Display *, Window, Window);
-void ewmh_root_set_current_desktop(Display *, Window, unsigned);
-void ewmh_root_set_number_of_desktops(Display *, Window, unsigned);
+void ewmh_root_set_current_desktop(Display *, Window, unsigned long);
+void ewmh_root_set_number_of_desktops(Display *, Window, unsigned long);
 void ewmh_root_set_desktop_names(Display *, Window, unsigned char *, unsigned);
 
 /* application window properties
